Struct ispell::SpellChecker

pub struct SpellChecker { /* fields omitted */ }

Spell Checker

Checks the spelling of a list of words

Example

use ispell::SpellLauncher;
let mut checker = SpellLauncher::new().launch().unwrap();
let errors = checker.check("This should not contain any error").unwrap();
assert!(errors.is_empty());Run

Methods

impl SpellChecker

fn check(&mut self, text: &str) -> Result<Vec<IspellError>>

Checks the spelling of a string

This method only returns the error that ispell detects.

fn check_raw(&mut self, text: &str) -> Result<Vec<IspellResult>>

Checks the spelling of a string

This method returns a vector of all ispell answers, even when there is no errors. Usually, the check method, which only returns errors, will be more useful.