Crate ispell

This library provides an interface for easily calling the ispell or aspell command from Rust programs.

Example

let mut checker = SpellLauncher::new()
                 .aspell(true)
                 .launch()
                 .unwrap();
let errors = checker.check("Testing iff if it works").unwrap();
assert_eq!(&errors[0].mispelled, "iff");Run

Warning

This library hasn't been very tested yet. It tries to avoid panic!s but, unfortunately, since Reads are blocking, there is a risk that it will simply hang up infinitely.

Structs

Error
Error type returned by methods of this library

IspellError
An ispell error when a word is not found in a dictionary

SpellChecker
Spell Checker

SpellLauncher
Spell Launcher wizard (ah, ah).

Enums

IspellResult
A result from ispell, corresponding to a line that is sent back for each word

Type Definitions

Result
Result type (returned by most methods of this library)